+++ packages/atomic/src/components/facets-v1/atomic-rating-facet/atomic-rating-facet.tsx
@@ -58,9 +58,8 @@ import {registerFacetToStore} from '../../../utils/store';
   shadow: true,
 })
 export class AtomicRatingFacet
-  implements
-    InitializableComponent,
-    BaseFacet<NumericFacet, NumericFacetState> {
+  implements InitializableComponent, BaseFacet<NumericFacet, NumericFacetState>
+{
   @InitializeBindings() public bindings!: Bindings;
   public facet!: NumericFacet;
   public searchStatus!: SearchStatus;