+++ packages/atomic/src/components/facets/atomic-facet/atomic-facet.tsx
@@ -63,7 +63,8 @@ import {registerFacetToStore} from '../../../utils/store';
   shadow: true,
 })
 export class AtomicFacet
-  implements InitializableComponent, FacetSearchComponent, BaseFacetState {
+  implements InitializableComponent, FacetSearchComponent, BaseFacetState
+{
   @InitializeBindings() public bindings!: Bindings;
   public facet!: Facet;
   public searchStatus!: SearchStatus;
@@ -117,7 +118,12 @@ export class AtomicFacet
    */
   @Prop() public enableFacetSearch = true;
   /**
-   * The sort criterion to apply to the returned facet values. Possible values are 'score', 'numeric', 'occurrences', and 'automatic'.
+   * The sort criterion to apply to the returned facet values. Possible values are:
+   *
+   * * 'score'
+   * * 'numeric'
+   * * 'occurrences'
+   * * 'automatic'
    */
   @Prop() public sortCriteria: FacetSortCriterion = 'automatic';