+++ packages/atomic/src/components/result-template-components-v1/atomic-result-fields-list/atomic-result-fields-list.tsx
@@ -25,7 +25,7 @@ export class AtomicResultFieldsList {
     this.resizeObserver.disconnect();
   }
 
-  public componentWasRendered() {
+  public componentDidRender() {
     this.update();
   }