+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.72.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.71.1...@coveo/atomic@0.72.0) (2021-08-27)
+
+
+### Features
+
+* **atomic:** used InteractiveResult controller with atomic-result-link ([#1139](https://github.com/coveo/ui-kit/issues/1139)) ([51af1e3](https://github.com/coveo/ui-kit/commit/51af1e3ab01fbb77c3ef7d2d7485e794de9c4e92))
+
+
+
+
+
+## [0.71.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.71.0...@coveo/atomic@0.71.1) (2021-08-27)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.71.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.70.2...@coveo/atomic@0.71.0) (2021-08-26)
+
+
+### Features
+
+* **atomic:** create v1 Pager & Result per Page Components  ([#1123](https://github.com/coveo/ui-kit/issues/1123)) ([6a88bb3](https://github.com/coveo/ui-kit/commit/6a88bb3085fab588f8709f5d5dc487a843509ba8))
+
+
+
+
+
+## [0.70.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.70.1...@coveo/atomic@0.70.2) (2021-08-26)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.70.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.70.0...@coveo/atomic@0.70.1) (2021-08-25)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.70.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.69.1...@coveo/atomic@0.70.0) (2021-08-25)
+
+
+### Features
+
+* **atomic:** new V1 atomic-breadbox component ([#1117](https://github.com/coveo/ui-kit/issues/1117)) ([a775987](https://github.com/coveo/ui-kit/commit/a7759878c21d4b534ab326671ce8bed8e7ce326c))
+
+
+
+
+
 ## [0.69.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.69.0...@coveo/atomic@0.69.1) (2021-08-23)