+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.18.0...@coveo/headless@1.18.1) (2021-08-27)
+
+
+### Bug Fixes
+
+* **headless:** fix analytics click event for result folding ([#1133](https://github.com/coveo/ui-kit/issues/1133)) ([0b25520](https://github.com/coveo/ui-kit/commit/0b2552097f0cc32a7c87e110f1ab9984315392ad))
+* **headless:** search parameter preventing auto selection on category facet ([#1127](https://github.com/coveo/ui-kit/issues/1127)) ([8ae636e](https://github.com/coveo/ui-kit/commit/8ae636eb631772ffc208f870926ed01036523b8e))
+
+
+
+
+
+# [1.18.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.17.3...@coveo/headless@1.18.0) (2021-08-26)
+
+
+### Features
+
+* **headless:** added recent results controller ([#1119](https://github.com/coveo/ui-kit/issues/1119)) ([7931d3c](https://github.com/coveo/ui-kit/commit/7931d3c774b468178e0093833935f634ff18455a))
+
+
+
+
+
+## [1.17.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.17.2...@coveo/headless@1.17.3) (2021-08-25)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+## [1.17.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.17.1...@coveo/headless@1.17.2) (2021-08-25)
+
+
+### Bug Fixes
+
+* **headless:** corrected actions mismatch and slice list ([#1125](https://github.com/coveo/ui-kit/issues/1125)) ([18eb86f](https://github.com/coveo/ui-kit/commit/18eb86f1763182ba840a2a7bef364620b7656f52))
+
+
+
+
+
 ## [1.17.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.17.0...@coveo/headless@1.17.1) (2021-08-23)